fix(ci): clippy -D warnings sur l'ensemble du workspace
CI échouait depuis plusieurs pushs sur cargo clippy (~95 erreurs accumulées).
Trois familles de fix :

1. Auto-fixes via cargo clippy --fix (60+ sites) :
   - useless conversions, manual_split_once, manual_range_contains,
     unnecessary casts/map_or/filter_map, redundant patterns, etc.

2. Allow ciblés sur du code volontairement structuré :
   - SOCKET (FFI Windows API) → upper_case_acronyms
   - MusicDiscKind (variants C418_*) → non_camel_case_types
   - dispatch_command_line / ExecutionContext::new / placement_y_for_footprint
     / attack_entity → too_many_arguments
   - command_selectors::SelectorKind::from_str → should_implement_trait
   - RawFunction / RawCondition (serde) → dead_code (champs non lus)
   - chunk_serializer biome coarse loop → needless_range_loop

3. Refactor manuels :
   - biome_color : range 11..=30 ordonnée après JUNGLE_COLORS (était unreachable)
   - block_wall::has_post : reformule les bool pour clippy::nonminimal_bool
   - snow_layer::snowballs_dropped : branches identiques → unifiées
   - pointed_dripstone/trade_offer : .max().min() → .clamp()
   - main.rs ServerSnapshot : construction directe sans Default+reassign
   - leaves_decay : match → !matches!() macro
   - empty_line_after_doc_comments dans 6 fichiers (amethyst_resonance,
     cocoa_beans, deepslate, lush_caves, tick_speed, trial_chamber)

cargo clippy -- -D warnings passe maintenant sans erreur.

+++ crates/mc-rs-server/src/binary_stream.rs
@@ -87,7 +87,7 @@ impl BinaryStream {
         let mut v = v as u32;
         loop {
             if v >= 0x80 {
-                self.write_byte((v as u8 | 0x80) as u8);
+                self.write_byte(v as u8 | 0x80);
                 v >>= 7;
             } else {
                 self.write_byte(v as u8);
